* Move keys func to pkg
